requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pandas
duckdb
matplotlib
kaggle
kagglehub
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
pytest
httpx